Firebase vuex actions moved to separate service and used from there
